backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
python-multipart
